Put back new type system  (#202)
This puts back the FloeCat type system refactor commit that was reverted recently.

Canonical logical kinds are enforced end‑to‑end.
Unknown/unmapped types fail fast in FloeCat (FloeDB retains its own BINARY fallback policy).
TIMESTAMP/TIMESTAMPTZ semantics are consistent across coercion/encoding/comparison/Arrow.
Non‑orderable types no longer emit stats; complex types reject min/max encoding.
INTERVAL is now modeled with range + precisions and encoded as ISO‑8601 in stats.
Arrow mapping is strict and aligned with canonical types; decimals use Decimal256 when needed.
Documentation and tests were updated to reflect the new contract.
---------

+++ service/src/test/java/ai/floedb/floecat/connector/dummy/DummyConnector.java
@@ -48,14 +48,14 @@ public final class DummyConnector implements FloecatConnector {
 
   private static final List<Col> LEAF_COLS =
       List.of(
-          new Col(1, "id", "INT64", 1),
+          new Col(1, "id", "INT", 1),
           new Col(2, "ts", "TIMESTAMP", 2),
           // struct user (3) -> id(4), name(5)
-          new Col(4, "user.id", "INT64", 3),
+          new Col(4, "user.id", "INT", 3),
           new Col(5, "user.name", "STRING", 4),
           // list items (6) -> element(7 struct) -> sku(8), qty(9)
           new Col(8, "items.element.sku", "STRING", 5),
-          new Col(9, "items.element.qty", "INT32", 6),
+          new Col(9, "items.element.qty", "INT", 6),
           // map attrs (10) -> key(11), value(12)
           new Col(11, "attrs.key", "STRING", 7),
           new Col(12, "attrs.value", "STRING", 8));
